#587: refactor booklist to support game and book
