--- packages/gdl-frontend/components/GameList/GameLink.js
@@ -1,84 +1,0 @@
-// @flow
-/**
- * Part of GDL gdl-frontend.
- * Copyright (C) 2017 GDL
- *
- * See LICENSE
- */
-
-import * as React from 'react';
-import { css } from '@emotion/core';
-import styled from '@emotion/styled';
-import { Card, CardContent, Typography } from '@material-ui/core';
-
-import { Link } from '../../routes';
-import CoverImage from '../CoverImage';
-import media from '../../style/media';
-
-export const coverWidths = {
-  small: 105,
-  large: 130
-};
-
-export type Book = $ReadOnly<{
-  id: string,
-  bookId: number,
-  title: string,
-  language: {
-    code: string
-  },
-  coverImage: ?{ url: string }
-}>;
-
-/**
- * Adds an absolute anchor above the whole cover, so you can click anywhere.
- * It is hidden from screen readers and when using the keyboard, in that case the title is also a link.
- */
-export default ({ game }: { game: Book }) => (
-  <Card css={cardCss}>
-    <Link route="game" params={{ id: game.id, lang: game.language }} passHref>
-      <ClickTarget aria-hidden tabIndex="-1" data-cy="game-link" />
-    </Link>
-    <CoverImage size="small" coverImage={game.coverImage} noShadow />
-    <CardContent css={{ padding: 10, ':last-child': { paddingBottom: 10 } }}>
-      <Link route="game" params={{ id: game.id, lang: game.language }} passHref>
-        <Typography
-          lang={game.language.code}
-          title={game.title}
-          noWrap
-          component="a"
-          align="center"
-        >
-          {game.title}
-        </Typography>
-      </Link>
-    </CardContent>
-  </Card>
-);
-
-/**
- * Add small brightness effect to book cover when hovered
- */
-const cardCss = css`
-  position: relative;
-  box-shadow: 0 10px 30px 0 rgba(0, 0, 0, 0.1);
-  &:hover {
-    img {
-      transition: 1s opacity linear;
-      pointer-events: none;
-      filter: opacity(0.9);
-    }
-  }
-  width: ${coverWidths.small}px;
-  ${media.tablet`
-    width: ${coverWidths.large}px;
-  `};
-`;
-
-const ClickTarget = styled('a')`
-  position: absolute;
-  left: 0;
-  right: 0;
-  top: 0;
-  bottom: 0;
-`;
