+++ packages/gdl-frontend/components/HomePage/index.js
@@ -13,6 +13,7 @@ import { Button, Card, CardContent, Typography } from '@material-ui/core';
 
 import type {
   Category,
+  Games_games as Game,
   BooksAndFeatured,
   BooksAndFeatured_featuredContent as FeaturedContent
 } from '../../gqlTypes';
@@ -27,8 +28,7 @@ import {
   CategoryNavigation
 } from '../../components/NavContextBar';
 import Head from '../../components/Head';
-import BookList from '../../components/BookList';
-import GameList from '../../components/GameList';
+import ScrollView from '../../components/ScrollView';
 import { colors, spacing } from '../../style/theme';
 import media from '../../style/media';
 import { flexCenter } from '../../style/flex';
@@ -80,6 +80,7 @@ const HeroCardTablet = styled(Card)`
 `;
 
 type Props = {|
+  games: Array<Game>,
   bookSummaries: $Diff<
     BooksAndFeatured,
     { featuredContent: Array<FeaturedContent> }
@@ -170,16 +171,16 @@ class HomePage extends React.Component<Props> {
             <CardContent>{cardContent}</CardContent>
           </HeroCardMobile>
 
-          <View {...bookListViewStyle}>
+          <View {...scrollListViewStyle}>
             <Container width="100%">
-              <BookList
+              <ScrollView
                 shouldBeColorized
                 heading={<Trans>New arrivals</Trans>}
                 browseLinkProps={{
                   lang: languageCode,
                   category: category
                 }}
-                books={NewArrivals.results}
+                items={NewArrivals.results}
               />
             </Container>
           </View>
@@ -191,9 +192,9 @@ class HomePage extends React.Component<Props> {
                 data.results && data.results.length > 0
             )
             .map(([level, data]: [ReadingLevel, any]) => (
-              <View {...bookListViewStyle} key={level}>
+              <View {...scrollListViewStyle} key={level}>
                 <Container width="100%">
-                  <BookList
+                  <ScrollView
                     heading={<ReadingLevelTrans readingLevel={level} />}
                     level={level}
                     shouldBeColorized
@@ -202,16 +203,21 @@ class HomePage extends React.Component<Props> {
                       readingLevel: level,
                       category: category
                     }}
-                    books={bookSummaries[level].results}
+                    items={bookSummaries[level].results}
                   />
                 </Container>
               </View>
             ))}
 
           {category === 'Library' && (
-            <View {...bookListViewStyle}>
+            <View {...scrollListViewStyle}>
               <Container width="100%">
-                <GameList games={games} />
+                <ScrollView
+                  items={games}
+                  shouldBeColorized
+                  level="Games"
+                  heading="Games"
+                />
               </Container>
             </View>
           )}
@@ -221,7 +227,7 @@ class HomePage extends React.Component<Props> {
   }
 }
 
-const bookListViewStyle = {
+const scrollListViewStyle = {
   py: spacing.medium,
   borderBottom: `solid 1px ${colors.base.grayLight}`
 };
