+++ packages/gdl-frontend/components/ScrollView/coverWidths.js
@@ -1,0 +1,12 @@
+// @flow
+/**
+ * Part of GDL gdl-frontend.
+ * Copyright (C) 2017 GDL
+ *
+ * See LICENSE
+ */
+
+export const coverWidths = {
+  small: 105,
+  large: 130
+};
