+++ packages/gdl-frontend/elements/index.js
@@ -14,3 +14,5 @@ export { default as DelayedLoading } from './DelayedLoading';
 export { default as IconButton } from './IconButton';
 export { default as Center } from './Center';
 export { default as Hidden } from './Hidden';
+export { default as Scroller } from './Scroller';
+export { default as Shimmer } from './Shimmer';
